Add Panasonic TX-P50C10E remote
I recorded the information from the official remote and everything works. Note that I could not test the CH+ and CH- in depth because I did not configure the analog TV. However, the buttons seem to react as they should. If you have any problem, don't hesitate to contact me !