add Fakir VC60 Fan, both 2019+2020 model
The Fakir VC60 fans manufactured in 2019 or before have different
remotes than the ones manufactured in 2020 or later.

The newer versions of this fan have a button to disable the (very
loud) beep on their remote, which the 2019 or earlier versions do
not have.

Sadly the newer remote doesn't seem to be able to disable the
beep on older models since they're quite different to one another.