Create Kenwood_DMX4704s.ir
IR Remote file that uses FlipperZero as infrared remote on this model of Kenwood head unit for car stereo.  Tested, and seems to work fine with my head unit.

The functionality is limited.  I had this trouble with an aftermarket remote I had purchased too.  I believe the infrared system on this head unit is very limited.  The user's manual also only shows the functions here as available on the factory optional remote.