Add Epson EBX7 Remote - Epson Controller 154720000
The numpad keys are named from 0 to 9 and then Num0 to Num9, some of those keys weren't sending a code while being pressed but did/sent a different code than the key alone when Num was pressed along with the key. So 0 to 9 means the key alone and Num0 to Num9 means Num+Key.