Add Epson EB-E20 / PowerLite E20
Adding Epson EB-E20 / PowerLite E20 cloned from the original remote, already tested. I'm adding this remote since none of the generic already available on the original repo seemed to work. Also, both the Epson EB-E20 and PowerLite E20 projectors use an infrared protocol that sends two consecutive codes when a remote control button is pressed and held: one for verification and one with the actual command. This behavior can make it difficult to accurately emulate commands with devices like the Flipper Zero, especially when using predefined protocols, which is why it is a raw remote.